# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyNet API",
    "description": "Project-owned API for the MyNet attendee workspace. Every endpoint except sign-in requires a valid sign-in session cookie, and every authenticated endpoint resolves the attendee from that cookie server-side — never from a client-supplied identifier (FR-035, FR-036).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mynet_session",
        "description": "Opaque sign-in session token. HttpOnly, Secure, SameSite=Lax. Never appears in a response body."
      }
    },
...
  },
  "paths": {
...
    "/events/{eventId}/attendees/{attendeeId}": {
      "get": {
        "summary": "Another attendee's profile, within a conference you share",
        "tags": [
          "profile"
        ],
        "description": "Carries both guards. The READER's registration is proven by the branded EventScope the route audit guarantees is present; the TARGET must additionally be registered for the same conference, be discoverable, and have a verified address — all three evaluated server-side before any field is returned (FR-390).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "attendeeId",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "attendeeId",
                    "displayName",
                    "interests",
                    "hasAvatar"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "attendeeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "company": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "role": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "headline": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "networkingIntent": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "enum": [
                        "open_to_meetings",
                        "open_to_messages",
                        "not_networking",
                        null
                      ]
                    },
                    "availability": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "enum": [
                        "available",
                        "busy",
                        null
                      ]
                    },
                    "interests": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "hasAvatar": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No conference in common, **or** no such attendee, **or** discoverability off, **or** the address unverified. All four are deliberately indistinguishable — identical status and identical body — so a requester cannot learn who exists, who is hiding, or whose address is verified (FR-361).",
            "content": {
              "application/json": {
                "schema": {
                  "description": "No conference in common, **or** no such attendee, **or** discoverability off, **or** the address unverified. All four are deliberately indistinguishable — identical status and identical body — so a requester cannot learn who exists, who is hiding, or whose address is verified (FR-361).",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/{eventId}/attendees/{attendeeId}/avatar": {
      "get": {
        "summary": "Another attendee's avatar, under the same conditions as their profile",
        "tags": [
          "profile"
        ],
        "description": "THE SAME THREE CONDITIONS as the profile, reusing the same query rather than repeating the predicate — serving a hidden attendee's photograph would be the profile withheld and the face given away. 204 when they have none, so the client renders the fallback (FR-351).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "eventId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "attendeeId",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "contentType",
                    "base64"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "contentType": {
                      "type": "string"
                    },
                    "base64": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No avatar. The fallback renders."
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No conference in common, **or** no such attendee, **or** discoverability off, **or** the address unverified. All four are deliberately indistinguishable — identical status and identical body — so a requester cannot learn who exists, who is hiding, or whose address is verified (FR-361).",
            "content": {
              "application/json": {
                "schema": {
                  "description": "No conference in common, **or** no such attendee, **or** discoverability off, **or** the address unverified. All four are deliberately indistinguishable — identical status and identical body — so a requester cannot learn who exists, who is hiding, or whose address is verified (FR-361).",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}